requirements.txt
streamlit==1.31.1
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
numpy==1.26.3
plotly==5.18.0
reportlab==4.0.4
sqlalchemy==2.0.39
